requirements.txt
python-dotenv>=0.19.0
python-telegram-bot>=20.0
